refactor initial leaderboard page
