+++ packages/client/src/components/organisms/app-routes/AppRoutes.tsx
@@ -2,7 +2,7 @@ import { Route, Routes } from 'react-router-dom'
 import { LandingPage } from '@pages/landing'
 import { AuthPage } from '@pages/auth'
 import { SignupPage } from '@pages/sign-up'
-import Leaderboard from '@pages/Leaderboard'
+import { LeaderboardPage } from '@pages/leaderboard'
 import { ProfilePage } from '@pages/profile'
 import { ErrorBoundary } from '@organisms/error-boundary'
 import { routes } from './routes'
@@ -20,7 +20,7 @@ const AppRoutes = () => {
         <Route path={routes.signUp.path} element={<SignupPage />} />
         <Route element={<ProtectedRoute />}>
           <Route path={routes.profile.path} element={<ProfilePage />} />
-          <Route path={routes.leaderboard.path} element={<Leaderboard />} />
+          <Route path={routes.leaderboard.path} element={<LeaderboardPage />} />
           <Route path={routes.main.path} element={<BeginGamePage />} />
           <Route path={routes.game.path} element={<GamePage />} />
           <Route path={routes.score.path} element={<EndGamePage />} />
