--- packages/client/src/components/pages/Leaderboard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Leaderboard'
