+++ packages/client/src/components/pages/leaderboard/Leaderboard.tsx
@@ -8,7 +8,7 @@ import {
   TableBody,
   Container, Box, Avatar, Typography
 } from '@mui/material'
-import mockData from '@pages/Leaderboard/mockData';
+import mockData from './mockData';
 import { ILeaderboard } from '@src/types/leaderboard'
 import { withNavbar } from '@services/withNavbar'
