+++ packages/client/src/components/pages/leaderboard/index.ts
@@ -1,0 +1,1 @@
+export { default as LeaderboardPage } from './Leaderboard'
